// src/app/models/Athlete.ts
export class Athlete {
  name: string;
  dob: Date;
  sports: string;
  nationality: string;
  gender: string;
  about: string;
  location: string;
  married: string;
  team: string;
  association: string;
  interests: string;
  charaties: string;
  pets: string;
  drinks: string;
  instagram: string;
  facebook: string;
  youtube: string;
  twitch: string;
  snapchat: string;
  twitter: string;
}